pyproject.toml
[tool.poetry]
name = "nicegui-chatbot"
version = "0.1.0"
description = "Multi-agent chat framework with NiceGUI, LangGraph, Postgres, and modular orchestration."
authors = ["Your Name <your@email.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
nicegui = "^1.4.20"
loguru = "^0.7.2"
SQLAlchemy = "^2.0.30"
asyncpg = "^0.29.0"
langgraph = "^0.0.10"
openai = "^1.30.1"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
Pillow = "^10.3.0"
starlette = "^0.36.3"
psycopg2-binary = "^2.9.9"
cryptography = "^42.0.5"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
flake8 = "^7.0.0"
black = "^24.4.2"
